[MOD] UserService interface
